+++ 1.21.6/Common/src/main/java/fuzs/puzzleslib/api/client/event/v1/gui/ScreenOpeningCallback.java
@@ -1,7 +1,6 @@
 package fuzs.puzzleslib.api.client.event.v1.gui;
 
 import fuzs.puzzleslib.api.event.v1.core.EventInvoker;
-import fuzs.puzzleslib.api.event.v1.core.EventResult;
 import fuzs.puzzleslib.api.event.v1.core.EventResultHolder;
 import net.minecraft.client.gui.screens.Screen;
 import org.jetbrains.annotations.Nullable;
@@ -15,10 +14,10 @@ public interface ScreenOpeningCallback {
      * {@link net.minecraft.client.Minecraft#setScreen}, allows for exchanging the new screen with a different one, or
      * can prevent a new screen from opening, by returning the original screen (which will be initialized once again).
      *
-     * @param oldScreen the screen that is being removed, may be null when opening the screen from
+     * @param oldScreen the screen that is being removed, which may be {@code null} when opening the screen from
      *                  {@link net.minecraft.client.gui.Gui}, like {@link net.minecraft.client.gui.screens.PauseScreen}
-     * @param newScreen the new screen that is being set, may be <code>null</code> when closing a screen and returning
-     *                  to the in-game gui
+     * @param newScreen the new screen that is being set, which may be {@code null} when closing a screen and returning
+     *                  to the in-game hud
      * @return <ul>
      *         <li>{@link EventResultHolder#interrupt(Object)} to set a different new screen, potentially the original screen for no change (except the original screen being initialized again), or <code>null</code> when closing a screen and returning to the in-game gui</li>
      *         <li>{@link EventResultHolder#pass()} to allow the vanilla screen to be set</li>